metrics: Allow for fractional values
The psuedo-standard spec for the statsd protocol says that for maximum
compatibility values should be integers, but most implementations do
actually support floating point values. Especially for timers with very
low magnitudes (in the 1-2ms range) rounding to the nearest integer
leads to very ugly graphs that are flipflopping between two values.