Reorganize installation of binaries
This makes it so that setup.py installs baseplate-{serve,healthcheck}
rather than only being installed when a debian package. The major
version of the Python interperter running setup.py will be added to the
names of the binaries to allow the library to be installed for Python 2
and 3 simultaneously, just like it was as a debian package.