docs: Reorganize and clarify
This stops hiding the module docs under the API section and pulls it to
the front, adding narrative docs to the module docs themselves. The idea
here is that by intermingling the two we can be more clear and less
redundant. This is closer to how the Python stdlib docs work too.
Additionally, the docs are now split up into CLI/Framework/Library
sections so that it's more obvious which parts are related to which
aspect of Baseplate.