Pass exception information to span observers when spans end
This allows us to take different actions depending on if the span exited
cleanly or exceptionally, such as automatically sending a "this request
failed" metric or the like.

Note: this is an API-breaking change. r2 is currently the only
subclasser of Span outside of baseplate, so that should be safe to fix
up before rolling this out.