metrics: Log metrics at debug level in NullTransport
This is useful during development to ensure that you're sending the
metrics you intend. The reason for doing this in the transport, rather
than in the individual metric classes is so that the logging only occurs
when the metrics are actually flushed so that any potential bugs related
to unflushed metrics aren't masked.