thrift_pool: Rework exception hierarchy
This eliminates the custom exceptions in thrift_pool and replaces them
with explicit raising of TTransportException. The goal is to make it so
that we can trust that TException-derived exceptions that are not also
derived from T{Application,Protocol,Transport}Exception are "user"
errors rather than transient system errors. This allows us to more
safely know when it's OK to retry as well as how to bubble exceptions
appropriately.