Add retry sugar to thrift context object client
This sugar allows for Thrift-specific (i.e. knowledge of exception
types) retries to be done on service clients. The reason for using a
context manager here is to hopefully reduce the chance that a copy of
the proxy client with special retry policy is saved for later and
accidentally is used for retries on non-idempotent operations.