Add SQLAlchemy diagnostics support
This adds a context factory for SQLAlchemy engines and ORM sessions. All
queries generated by a service using this factory will be tracked by the
diagnostics system and have trace information added to the query
statements.

This version uses a threadlocal (or greenlet-local when monkeypatched
gevent) to pass data to the SQLAlchemy events. I'm not at all thrilled
with this solution since baseplate tries so hard to avoid globalish
state.