Rework diagnostics observer registration
Since we have access to the root and child spans in the observer
callbacks, there's no need for the special return value logic. This
removes that special logic and lets observers just call the register()
method on spans. To support this, the register() method is now public
and documented properly. The goal here is just to simplify and make
the API a bit more explicit.