Centralize and standardize creation of trace info
Previously, the framework integrations decided how trace IDs were
validated and generated. This meant there was no guaranteed consistency.

This makes it so that we can assume in downstream code that a trace_id
etc. is available during request processing and that it has a common
format, no matter which framework we're using.