Event: Refactor serialization and payload handling
The event pipeline flattens the payload before ingest so having
entries in multiple sections isn't valid.

From the docs:
  These fields will be output at the same depth as everything in
  the payload, but will be obfuscated before output.

As a result of this the `obfuscated` parameter has been removed
from Event#get_field, and the `obfuscate` parameter has been
deprecated on Event#set_field in favor of specifying a `kind`.