Rework diagnostics/span APIs for OpenTracing compatibility
This reworks the vocabulary of the diagnostics/span APIs to match the
terminology of OpenTracing. OpenTracing is a standard set of terms and
APIs for distributed tracing that doesn't rely on a specific tracing
implementation. This also separates annotations into tags and logs,
matching the terminology OpenTracing uses.

This does _not_ implement baggage right now as we don't have a use for
it at the moment, but it should be simple enough to add later.

For more info: http://opentracing.io/