Fix server span error catching for Pyramid integration
Previously, we didn't actually catch errors for the server span under
Pyramid because the NewResponse event is after the exception tween has
caught and morphed the exception into an error page. This uses a tween
to catch the exception and put its exc_info into the server span's
finish() event as expected.