thrift: Fix binary protocol responses from header protocol server
THeaderTransport detects the protocol of the client and can accept raw
binary protocol (without headers). It's supposed to then use the same
protocol when responding. Because we were using a different instance of
THeaderTransport for the input and output streams, the detected format
was not being used when responding. This lead to binary protocol clients
hanging as they got invalid data in response. This resolves the issue by
using the same transport and protocol instances in both directions.