Ensure thrift server spans get proper exception context
This makes it possible to rely on the exception context on thrift server
spans. handlerError is called in the generated code inside of the
except: branch of processor code. It will only be called for unexpected
exceptions. handlerDone is called either way immediately after but in
the case where handlerError was already called, the result will be a
TApplicationException marking the unexpected error.