Use where= instead of tests= in setup.cfg
`where` tells nose where to find the tests, while `tests` specifies the
tests to run. While both worked for the common case of running the whole
test suite, using the `tests` option made it so it was impossible to
specify a subset of tests to run on the command line.