Authentication context propagation (#120)
Adds wrappers and handlers for retrieving propagated authentication
information from upstream requests into an easy to use interface and
standard system for further downstream propagation.  Overall the process
is to convert the JWT payload received from the authentication service
into an `AuthenticationContext` which can do validation and extraction,
then store this is the current context and downstream requests via the
built in thrift context will pull this value out of the current context
and store the necessary information in a standard way for downstream
services to easily and automatically retrieve.

There are some assumptions that may need to be refined, such as the way
that the public RSA key for the authentication service is stored and
retrieved, currently it is assumed (and hardcoded) to be in a
`SecretStore` object under the simple key of
`jwt/authentication/secret`.  Also, it is currently dependent on the
originating service of the authentication to correctly package and
define the information in the ongoing context for retrieval in the
context handlers.  Any HTTP requests to downstream services will require
the correct header be defined by hand (as there is no baseplate based
HTTP client).