Encode AuthenticationContext token
Thrift expects a header to be a str in Python 2 (str or bytes works in
Python 3) so we `encode` the authentication token in the
`AuthenticationContext` constructor so `AuthenticationContext._token` is
always `bytes` in Python 3 or `str` in Python 2.

If we do not do this, the value may be `unicode` in Python 2.  `unicode` works
fine as long as the length of the string is under 128 characters but once you
go over that, you start to run into errors when Thrift tries to construct the
header.