Don't retry publishing forever when v2 event fails to validate
The previous check didn't work properly because requests.Response has
__bool__ defined as "Returns True if `status_code` is less than 400."
which meant that it was always falsey in the situations we wanted to
check it. Instead, we now just explicitly check if it's `None` and crash
on validation failure. This will discard the current message batch and
upstart will restart us to carry on with further messages.