Add thrift build step to setup.py
Building Baseplate strictly via `sdist` or other mechanisms
that rely on setup.py will lead to silent failures upon installation
for the baseplate.thrift module because there is no definition
in the setup.py build step for generating the included Baseplate
Thrift definitions in `baseplate/thrift/baseplate.thrift`. This adds
an override to build_py to generate Thrift files via the Makefile
and checks for a valid Thrift compiler.