Close thrift connection after unexpected exception in thrift
In cases where Thrift fails dealing with serialization/deserialization
(such as catastrophically wrong types being passed in) the connection
will end up in a bad state but we'll check it back into the pool to be
reused. This makes it so we close the connection in that case to prevent
it from causing issues for the next user.