Allow you to run the secret fetcher only once
While you generally want to run the secret fetcher as a daemon
so it will pick up any updates, there are some situations where
you only want to run it once.  This PR adds a `--once` argument
to `secret_fetcher.py` that allows you to run it a single time
and then exit.