Ensure that connection failures do not deplete ThriftConnectionPools. (#172)
This commit modifies `ThriftConnectionPool._acquire` in order to put
a blank (`None`) connection entry back in pool if the retry policy is
exhausted while trying to connect to the host. Previously, one
connection pool slot would be lost each time `_acquire` failed to
secure a new connection to the host.