Add ignore_exception_cls arg to configure_error_reporting
Some Python frameworks allow you to raise certain exceptions to indicate
certain errors, and the framework will handle that accordingly. Those type of
exceptions are raised intentionally and reporting them to sentry would be
noisy.

Add ignore_exception_cls arg to configure_error_reporting to enable the feature
of ignoring certain exceptions.

There are in general two approaches to this problem, one is passing in an
exception class and the other is passing in a function to test against
exc_info. The latter is more flexible (it can handle the case that there are
more than one exception classes to ignore that doesn't share the same
superclass), but the former is easier and can handle 90% of the cases.