Add BaseplateHeaderTrustHandler class to BaseplateConfigurator
In order to allow trace headers to be accepted between services we allow BaseplateConfigurator to be setup with a BaseplateHeaderTrustHandler. The default implementation provided here validates a header X-Trace-Hmac against a secret provided at the path secret/trace/hmac. If this validation passes then the trace headers for the request will be trusted.