secret fetcher: Generate our own nonce.
Vault stopped returning the nonce value in their response unless it
generated a nonce for the client in 0.9.0. The changelog also states
that the nonce will be logged if it's returned in the response.
Generating our own nonce will remove it from any logs.

N.B. Although this behavior change occured in 0.9.0, I'm updating the
test suite to run against the current 1.0 version.