Make ThriftContextFactory more efficient
The previous implementation would re-enumerate all the methods and
re-create all the proxy methods every time it was instantiated,
including on server spans, local spans, and even the retrying()
decorator. This method does the enumeration at factory instantiation and
uses the cached type.

Some unit tests were removed as they were overly knowledgeable about
internals and the new integration tests cover that area.

This is a port to the Apache Thrift version of Baseplate based on
@jesseyli's work in a fork.