Add 'TokenCSRFStoragePolicy' to pyramid integration
Rather that storing the token in another place and comparing
the stored token to what is returned by the client, this simply
checks that the token that is sent matches what we would expect
based on the request.

This only is available if the installed version of pyramid has
the `ICSRFStoragePolicy` interface available (it was added in
version 1.9).