Add 'CredentialSecret' secret type (#267)
Adds a new secret type for storing username/password pairs under the same key in Vault.  

This allows both the username and password to be updated simultaneously without having to juggle multiple vault keys, rolling your own way to encode the username and password into a single secret value, or trying to sync a username that exists as plaintext in a config file with an update to a secret in vault.