Monkeypatch ResponseFuture.result() to wait for callbacks
Add a callback_fn and errback_fn that are given special treatment: they must be
complete before a result will be returned from cassandra.cluster.ResponseFuture.result().
They are not given precedence over other callbacks or errbacks, so if
another callback triggers the response from the service (and the server
span is closed) the special callback might not complete.

This fixes a race condition where the server span can complete before
the callback has closed out the child span.