Rework context object to allow lazy attributes
This makes context attributes lazily get instantiated on first use each
request. Microbenchmarks show that this is a significant performance
improvement when the application doesn't use all of its context
attributes each request, but even when it does it's of comparable speed
so there's not much downside. However, there is a small impact in the
speed of access to an individual attribute once instantiated.