Simplify the request context laziness setup
This was jumping through too many hoops to not touch the wrapped pyramid
request object and it ended up not even working. It turns out that
pyramid's add_request_method with reify=True would try to use the
wrapped object directly and fail. So, this revives the mixin approach
for Pyramid (and just uses the class directly for Thrift) and is so able
to drastically simplify the laziness setup. We now get all the speed
benefit of context_config and its laziness but without the cost of
attribute accesses being proxied through __getattribute__ logic. And
pyramid works.