Reorganize modules and some of the code within them
This is a biggggg reshuffle and will require a lot of updates in
applications which we will do with automatic tooling.

Some of these are simple renames (context -> clients, integrations ->
frameworks, / -> lib), but there are also some subtler shifts like
moving the core stuff up to root `baseplate` and pulling edge context
out into its own lib.

The overall goal is to make the layout of the library much more obvious
to newcomers and to help improve the cleanliness of imports in
applications.