Default to 10 second wait for graceful server shutdown
The default was previously 0. I thought that meant wait forever, but
instead it meant die immediately. After fixing graceful shutdown, this
meant we'd still be hard-killing in flight requests but now we'd also be
getting the GreenletExit errors from it. This default of 10 seconds is
meant to give some time for long-running requests while not holding up
shutdown too long.