Fix infinite loop in Raven setup when using relative paths (#374)
When running a script that uses error_reporter_from_config() from a
directory that does not have any git repository, the revision resolution
would infinite loop. This grabs the absolute path to the code before
traversing the tree so that we can truly find out the revision or give
up.