Make BaseplateRequest play more nicely with superclass methods
Things like Request.from_file call the current subclass's constructor
with a specific order of arguments.[1] This doesn't work if we put our
context config at the front and make it required.

By reordering like this we play nicely as a subclass and can still get
our data in when we need it.

[1]: https://github.com/Pylons/webob/blob/ac2d238f50bed7fbbf115818ab7f208b71d7bdca/src/webob/request.py#L1251-L1253