Add list of exceptions for raven to always filter out (#517)
This change provides a default list  of exceptions for raven to filter out. In order to take advantage of the filtering you must not provide `sentry.ignore_exceptions` configuration. If you would like to provide exceptions to filter out in addition to the default list, you can provide a list of exceptions in the configuration variable `sentry.addtional_ignore_exceptions`.