Rework tagged metrics to ensure proper aggregations
Including "success" and "error" in the tags for meant we could never get
proper percentile aggregations across successful+failed requests.
Additionally, timers are often sampled more agressively than counters in
other implementations so we separate out the rate and the latency into
separate metrics to ensure accurate numbers.

This drops support for counting the rate of errors by exception name as
well as counting the rate of server timeouts.