Remove unused metric
For rediscluster we can't easily calculate the size of the pool since
counting the number of items in the queue doesn't work (the queue is
always full, just full of Nones if the queue is "empty").

We just decide to ignore that metric for now and just report the ones
that are easily available to us.