Add hot key tracker to the cluster reddis client
Add a feature to the cluster redis client that can enable hot key
tracking based on a configurable sample rate.

When read and/or write commands access keys in Redis we will sample
those commands and store the keys within a sorted set in Redis, which we
can use to check relative frequency of accessing some keys compared to
others.