fixes to thrift client metrics
* set protocol on thrift client

fixing observer ignoring data

* ensure span is started before handling request

when an exception was encontered before the span was started and
additional exception was raised by the `_call_thrift_method` exception
handler, as it expected the span to have been started.

this changes the semantics of the latency graph, as it now includes
every retry of a request!