[BASE-127] Fix thrift client connection reuse (introduced in 2.4.13) (#728)
Emit prometheus metrics for individual retry calls instead of whole
"transaction"
Do not emit metrics if the connection cannot be established to begin
with. (self.pool.connection() raises an exception)

The documentation for thrift_pool.py:ThriftConnectionPool.connection() reads:
> When the context is exited, the connection is returned to the pool.
> However, if it was exited via an unexpected Thrift exception, the
> connection is closed instead because the state of the connection is
> unknown.

As of 2.4.13, TTransportException (used to decide whether we should
retry or not) is raised, but caught before the context manager handling
pool.connection() can catch it. Which means that on retries, the
connections would end up in an unknown state, returned to the pool, and
yield potential bad results on subsequent uses. This change fixes that.