fix(RedisClusterClient): Add `redis_client_name` attribute (#743)
This updates RedisClusterClient.

Currently clients that use the `client_name` property to set the Redis
client with a Envoy Redis proxy are not able to connect to Redis. At
it's root this is an issue an issue with the redis-py library (see
redis/redis-py#2384).

Currently Baseplate's wrapper around the Redis client uses the
`client_name` to set the `redis_client_name` label on Prometheus metrics.
Due to the issue mentioned above clients who use a Redis Proxy are not
about to set the `redis_client_name` Prometheus label. The additional
`redis_client_name` property is now used to set the Prometheus label.
For backwards compatibility, if `client_name` is set, but
`redis_client_name` is not, the `redis_client_name` defaults to
`client_name`.