feat(prometheus): Run Prometheus server by default (#737)
The Prometheus server will run by default. The `metrics.enabled` configuration flag was added to disable to the Prometheus server if it shouldn't run.

This allows Baseplate users to disable statsd metrics without disabling the Prometheus server. In previous versions of Baseplate, if statsd was not configured the Prometheus server would not run.