test(RedisClient): Use common setup/teardown logic in tests (#744)
Currently not all cleanup tasks are done to ensure that Redis
integration tests can't interfere with each other. In particular I ran
into a case where adding a test would caused the `lock_test` to hang
when run a second time. This was because my new test happened to set a
key with the same name as the lock. To avoid this issue This change will
result in Redis clearing all data between each test run.

In addition this change clears all Redis Prometheus metrics between
every run.