fix(runtime metrics): Use a null statsd client for runtime metrics (#751)
Currently runtime metrics aren't being reported to Prometheus when StatsD is disabled. This change initialize a null StatD client to report StatsD metrics to.

This allows StatD metrics to continue to be reported for clients for clients that still depend on the metrics and also allows for clients that don't currently use Statsd to get runtime metrics in Prometheus.