Populate Cassandra cluster name from cluster metadata
The cluster name label allows us to map the client queries back to the
cluster they are being executed on. It can be used to determine client
query error rates and latencies broken down by cluster and map these
back to specific queries similar to how `cassandra_client_name` is
used.

If the metadata field is not populated, this returns an empty string.
The metadata field is populated on connect and refreshed at a regular
interval, however, the cluster name will not change.