Adds support for `MultiProcessValue` ids based on env variable
When baseplate is running as part of Monoceros process restarts
create too many files which make prometheus client extremely slow
and eventually it timeouts.

We want to be able to provide an alternative id to prometheus client
now based on an environment variable. To achieve this we set the
process id to:

* `MULTIPROCESS_WORKER_ID` if provided and fallback to PID in case
it doesn't exist
* Moved the prometheus client initialization before we make the server.

Signed-off-by: Sotiris Nanopoulos <sotiris.nanopoulos@reddit.com>