upgrade(actions): upgrade checkout and setup-python (#788) (#791)
GitHub deprecated Node 12 to Node 16. The current versions of checkout and setup-python GitHub actions depend on a depricated version of Node.

I cut a release as a test [here](https://github.com/reddit/baseplate.py/actions/runs/4985027857), the build failed because of a bad tag, but did run. [Here](https://github.com/reddit/baseplate.py/actions/runs/4963056617) is the previous failure.

This also upgrades pins the version of requests being used during build time, this is necessary for the package to be successfully uploaded.

#787 applies this change to the gcp build.